backend/app/services/agents.py
try:
    from crewai import Agent, Task, Crew, Process
except ImportError:
    # Mocking classes for dev without heavy deps
    class Agent:
        def __init__(self, **kwargs): pass
    class Task:
        def __init__(self, **kwargs): pass
    class Crew:
        def __init__(self, **kwargs): pass
        def kickoff(self): return "Mock Result"
    class Process:
        sequential = "sequential"

# Mock Agents Definition
def create_discovery_crew(problem_statement: str):
    # Definition of Agents
    literature_agent = Agent(
        role='Literature Researcher',
        goal='Analyze 5000+ papers for drug targets related to query',
        backstory='Expert textual analyst for biomedical literature.',
        verbose=True,
        allow_delegation=False
        # tools=[SearchTool()]
    )

    safety_agent = Agent(
        role='Safety Officer',
        goal='Validate candidates against ToxCast and FDA databases',
        backstory='Former regulator ensuring safety compliance.',
        verbose=True,
        allow_delegation=False
    )
    
    # Definition of Tasks
    task1 = Task(
        description=f'Find potential drug candidates for: {problem_statement}',
        agent=literature_agent
    )

    task2 = Task(
        description='Validate safety profile of identified candidates',
        agent=safety_agent
    )
    
    # Crew Orchestration
    crew = Crew(
        agents=[literature_agent, safety_agent],
        tasks=[task1, task2],
        verbose=2, 
        process=Process.sequential
    )
    
    return crew

async def run_crew_job(problem_statement: str):
    # In production, this runs in a Celery worker or separate container
    # crew = create_discovery_crew(problem_statement)
    # result = crew.kickoff()
    
    # Mock result for now
    return "Identified Metformin as a potential candidate."
