backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
alembic==1.13.1
asyncpg==0.29.0
pydantic==2.6.1
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
prometheus-client==0.19.0
reportlab==4.1.0
openpyxl==3.1.2
pandas==2.2.0
pytest==8.0.0
httpx==0.27.0
aiosqlite==0.19.0
